refactor directory structure #major
