+++ .github/workflows/push.yml
@@ -51,12 +51,26 @@ jobs:
         os: [ ubuntu-latest, macos-latest ]
     runs-on: ${{ matrix.os }}
     steps:
+      - name: Checkout code
+        uses: actions/checkout@v2
       - name: Install Go
         uses: actions/setup-go@v2
         with:
           go-version: 1.17.x
-      - name: Checkout code
-        uses: actions/checkout@v2
+      - name: Install Nginx
+        run: |
+          if [ "$RUNNER_OS" == "Linux" ]; then
+            sudo apt update
+            sudo apt install nginx
+            sudo nginx
+          elif [ "$RUNNER_OS" == "macOS" ]; then
+            brew install nginx
+            sudo nginx
+          else
+            echo "$RUNNER_OS not supported"
+            exit 1
+          fi
+        shell: bash
       - name: Test
         run: make test
 
@@ -95,11 +109,25 @@ jobs:
     steps:
       - uses: actions/checkout@v2
         with:
-          fetch-depth: 0  # Shallow clones should be disabled for a better relevancy of analysis
+          fetch-depth: 0
       - name: Install Go
         uses: actions/setup-go@v2
         with:
           go-version: 1.17.x
+      - name: Install Nginx
+        run: |
+          if [ "$RUNNER_OS" == "Linux" ]; then
+            sudo apt update
+            sudo apt install nginx
+            sudo nginx
+          elif [ "$RUNNER_OS" == "macOS" ]; then
+            brew install nginx
+            sudo nginx
+          else
+            echo "$RUNNER_OS not supported"
+            exit 1
+          fi
+        shell: bash
       - name: Coverage Test
         run: go test ./... -race -coverprofile=coverage.txt -covermode=atomic
       - name: SonarCloud Scan
