+++ internal/k8s/types/nodeport.go
@@ -1,0 +1,22 @@
+package types
+
+type NodePort struct {
+	MasterIP string
+	Port     int32
+	Workers  []*Worker
+}
+
+func NewNodePort(masterIP string, port int32) *NodePort {
+	return &NodePort{
+		MasterIP: masterIP,
+		Workers:  make([]*Worker, 0),
+		Port:     port,
+	}
+}
+
+// Equals method checks the equivalent of nodePort structs
+func (nodePort *NodePort) Equals(other *NodePort) bool {
+	isMasterIPEquals := nodePort.MasterIP == other.MasterIP
+	isPortEquals := nodePort.Port == other.Port
+	return isMasterIPEquals && isPortEquals
+}
